Move azure specific utils to AzureUtils.java
