+++ service/src/main/java/bio/terra/workspace/service/resource/controlled/cloud/azure/vm/CreateAzureVmStep.java
@@ -9,7 +9,7 @@ import bio.terra.stairway.exception.RetryException;
 import bio.terra.workspace.app.configuration.external.AzureConfiguration;
 import bio.terra.workspace.common.exception.AzureManagementException;
 import bio.terra.workspace.common.exception.AzureManagementExceptionUtils;
-import bio.terra.workspace.common.utils.AzureVmUtils;
+import bio.terra.workspace.common.utils.AzureUtils;
 import bio.terra.workspace.common.utils.FlightUtils;
 import bio.terra.workspace.db.ResourceDao;
 import bio.terra.workspace.generated.model.ApiAzureVmCreationParameters;
@@ -140,7 +140,7 @@ public class CreateAzureVmStep implements Step {
                       .setNetwork(networkInterface.primaryIPConfiguration().getNetwork())
                       .setSubnetName(subnetName)
                       .setDisk(existingAzureDisk.orElse(null))
-                      .setImage(AzureVmUtils.getImageData(creationParameters.getVmImage()))
+                      .setImage(AzureUtils.getVmImageData(creationParameters.getVmImage()))
                       .build()));
 
       context.getWorkingMap().put(AzureVmHelper.WORKING_MAP_VM_ID, createdVm.id());
@@ -228,13 +228,13 @@ public class CreateAzureVmStep implements Step {
               .withType(creationParameters.getCustomScriptExtension().getType())
               .withVersion(creationParameters.getCustomScriptExtension().getVersion())
               .withPublicSettings(
-                  AzureVmUtils.settingsFrom(
+                  AzureUtils.vmSettingsFrom(
                       creationParameters.getCustomScriptExtension().getPublicSettings()))
               .withProtectedSettings(
-                  AzureVmUtils.settingsFrom(
+                  AzureUtils.vmSettingsFrom(
                       creationParameters.getCustomScriptExtension().getProtectedSettings()))
               .withTags(
-                  AzureVmUtils.tagsFrom(creationParameters.getCustomScriptExtension().getTags()));
+                  AzureUtils.vmTagsFrom(creationParameters.getCustomScriptExtension().getTags()));
 
       if (creationParameters.getCustomScriptExtension().isMinorVersionAutoUpgrade()) {
         customScriptExtension.withMinorVersionAutoUpgrade();
